refactor: extract as BlockProviderV2
